requirements.txt
beautifulsoup4==4.9.2
lxml==4.5.2
python-telegram-bot==12.8
requests==2.24.0
